+++ docs/changelog.md
@@ -7,5 +7,6 @@
 * `build-code-coverage-metrics` - covert PHPUnit clover.xml into Markdown and badges
 * `build-contracts` - convert PHPUnit testdox.txt into Markdown
 * `build-php-badges` - get PHP badges from shields.io
+* `dump-config` - output the config that bengi is actually using
 * `bengi.json` config file for your project
 * `$HOME/.bengi.json` config file for your computer